Added Default Brightness. Other tweaks.
* Updated brightness status code to accurately reflect a 1 percent
brightness level (rather than rounding up to 2 percent).
* Added Default Brightness property for Hue Bulbs devices so their
features are more consistent with other lighting devices in Indigo
(such as SwitchLinc and LampLinc Dimmers).
* Changed "on", "off", and "toggle" action control functions so that
sending an "on" command to a Hue Bulb from within Indigo returns the
bulb to its previous brightness level, or its default brightness level
(if set).
* Updated Brightness, HSB, and Color Temperature setting methods so
that they properly save the brightness level specified in those actions
to the Hue Bulb's device properties for recall should an "on" command
be sent to it.
* Updated the bulb status gathering method so that if the bulb
brightness changes outside of the Hue Lights plugin (or the hub updates
the bulb brightness during a long transition time), Hue Lights does not
save the new brightness to the bulb properties and thus causing an "on"
command later to recall an incorrect previous brightness state.
* Changed the logging slightly to more closely match the log format of
native INSTEON device changes.