Added support for the Phoenix wall light.
Added support for the Phoenix wall light.  Note that the Phoenix light
will appear in the Hue bulb list.  This means that the plugin will
allow HSB or XY (RGB) color commands to be sent to the Phoenix even
though the Phoenix does not support such commands.  This will generate
an error from the Hue hub (and in the plugin).  Fuller support for
color-temperature-only lights will require more significant changes to
the plugin code later on.