Added new actions to the "Hue Group" device type:
  - advanceGroupScene
  - retreatGroupScene

When a Hue Group is added to Indigo that corresponds with a group defined in the Hue Hub,
these actions can be executed against the Indigo-defined group.

The actions get all of the scenes in the Hue Hub that are assigned to the group, and cycles
from them in advancing or retreating order every time the respective action is invoked.

Of course, this requires that you have Hue scenes in your hub, such as those added by
default with the iPhone Hue application.

In essence, every time advanceGroupScene is invoked, the next Hue Hub scene is applied to
that group's Hue lights. A sample use case might be triggers from a light switch, wherein
successive hitting of "ON" changes scenes until the user is satisfied with the appearance
of the lights (I use Insteon light switches in this manner with the modified plugin to
control my Hue lights).